Add support for DS9097 Thermometers through digitemp
This adds support for a new thermometer-type based on the DS9097 1Wire to Serial adaptor. This adaptor
is often sold in a package with a Serial to USB adaptor.

For reading from this device, the digitemp program — which is included in raspbian — is use. It can
read a variety of 1W-Thermometers and thus greatly extends the capabilities of CraftBeerPI.

Automatic discovery on the bus, as provided by digitemp, is used.