Add preliminary support for PiLight devices
PiLight is a multi-purpose home-automation framework to control
proprietary-RF devices. It can work on various bands like 433Mhz, 868
mHz and supports a variety of protcols.

For Brewing, the interesting part are cheap RF-power-plugs, like the
Intertechno 3500, which is a "kaku_switch" — or many others.

This adds support for talking to a PiLight-Server as an actor, so we can
control kettle-heatings or similar actuated with a PiLight-supported
socket or device. PiLight needs to be configured to listen on a static
port (it randomly chooses one by default), as well as expose devices
from its config-file. These are then auto-discovered and can be selected
in CraftBeerPi.